Feat: Add ending of match (#60)
* Feat: Add ending of match
This feature fully adds game ending logic. The win condition is determined on
every unit death, if player's unit was last, then the match is over, other
player is victorious (note: works only for 2-player gameplay). It is worth
mentioning that interfaces for pausing and resuming a match are implemented.

New files are introduced:
- end_game_modal.tscn — end-of-game modal, which displays which player won the
  match and allows you to go back to the match selection scene
- match_ui.gd (`MatchUi`) — core of Match UI
- darkened_blur.gdshader — shader for darkening and blurring a background (based
  on a great shader: https://godotshaders.com/shader/darkened-blur/)